seperate out query_qb as a standablone module- this way, can query a single entry or enter into loop. also add __init__
