+++ index_identifiers.py
@@ -3,32 +3,11 @@ from xml.etree import cElementTree as ET
 import time
 import csv
 import sys
+import query_db
+
 input_list=sys.argv[2]
 pause_time=int(sys.argv[1]) #5
 db=sys.argv[3]
-
-def query_db(db,id):
-    output=""
-    webenv_url='https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi?db='+db+'&term='+id+'&usehistory=y'
-    
-    try:
-        resp = requests.get(webenv_url)
-        root=ET.XML(resp.text)
-        we=root.find('WebEnv')
-        query_wenv=we.text
-        query_url="https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db="+db+"&webenv="+query_wenv+"&query_key=1"
-        ## Buried in this URL is the gene's symbol. FInd it. 
-        qeury_resp = requests.get(query_url)
-        #soup = BeautifulSoup(qeury_resp.content, 'html.parser')
-        # gene name start and stop
-        tmp=qeury_resp.text.replace("\n","")
-        start=tmp.find('locus "') + 7 # 861
-        end=tmp.find(",",start) - 1
-        output=tmp[start:end]
-    except:
-        output="ERROR"
-    out_line="NCBIGENE:"+id+"\t"+output
-    return(out_line)
 
 
 with open(input_list) as input_list_ids:
@@ -36,5 +15,5 @@ with open(input_list) as input_list_ids:
     for row in ids:
         id=row[0]
         time.sleep(pause_time)
-        print(query_db(db,id))
+        print(query_db.query_db(db,id))
